Add api_ modules for API views and URLs
